\name{RSearchYJ-package}
\alias{RSearchYJ-package}
\alias{RSearchYJ}
\docType{package}
\title{
RSearchYJ
}
\description{
  Web and Blog Search with Yahoo! Japan. (http://developer.yahoo.co.jp/webapi/search/)
}

\details{
\tabular{ll}{
Package: \tab RSearchYJ\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2010-06-26\cr
License: \tab BSD\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yohei Sato

Maintainer: Yohei Sato <yokkuns@tkul.jp>
}
\references{

}

\keyword{ package }
\seealso{
  
}
\examples{
#res.web <- yahooWebSearch("yahoo appid", "Yahoo !")
#res.web
#res.blog <- yahooBlogSearch("yahoo appid", "Yahoo !")
#res.blog
}
